refactor(mobile): split screens into dedicated screen and component files
Extract inline UI and sub-components out of Expo Router route files into
a clear two-layer structure:

- Route files (app/) now own only state, data-fetching, and navigation
  callbacks; all rendering is delegated to a named Screen component
- Screen components (src/screens/) own layout and orchestration for a
  given route, accepting data and handler props
- Reusable UI pieces are broken into focused components under
  src/components/: AuthHeader, SignInForm, SignUpForm, ClubCard,
  EmptyClubsState, and shared authFormStyles
- Auth stack gains a `fade` transition via _layout.tsx

No behaviour changes intended.
