+++ src/app/(auth)/_layout.tsx
@@ -1,5 +1,5 @@
 import { Stack } from "expo-router"
 
 export default function AuthLayout() {
-  return <Stack screenOptions={{ headerShown: false }} />
+  return <Stack screenOptions={{ headerShown: false, animation: "fade" }} />
 }
