+++ src/app/(auth)/sign-up.tsx
@@ -1,34 +1,17 @@
 import { useState } from "react"
-import { View, ActivityIndicator, StyleSheet } from "react-native"
 import { router } from "expo-router"
 
-import { Screen } from "@/components/Screen"
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Text } from "@/components/ui/text"
+import { SignUpScreen } from "@/screens/SignUpScreen"
 import { api } from "@/services/api"
 import { useAuthStore } from "@/stores/authStore"
-import { useAppTheme } from "@/theme/context"
-
-const inputClassName =
-  "bg-surface-container border-outline-variant/50 rounded-sm h-[50px] px-[14px] text-sm text-on-surface"
-
-const labelClassName = "text-xs uppercase tracking-[0.1em] text-on-surface"
 
 export default function SignUp() {
-  const { theme } = useAppTheme()
-  const colors = theme.colors
-  const [name, setName] = useState("")
-  const [email, setEmail] = useState("")
-  const [password, setPassword] = useState("")
-  const [confirm, setConfirm] = useState("")
   const [loading, setLoading] = useState(false)
   const [error, setError] = useState<string | null>(null)
   const [fieldErrors, setFieldErrors] = useState<Record<string, string[]>>({})
   const setTokens = useAuthStore((state) => state.setTokens)
 
-  async function handleSignUp() {
+  async function handleSignUp(name: string, email: string, password: string) {
     setLoading(true)
     setError(null)
     setFieldErrors({})
@@ -49,125 +32,12 @@ export default function SignUp() {
   }
 
   return (
-    <Screen preset="scroll" safeAreaEdges={["top"]} contentContainerStyle={styles.content}>
-      <Text className="text-center text-3xl text-on-surface mt-14" style={styles.titleFont}>
-        NextChapter
-      </Text>
-      <Text className="text-center text-sm text-outline mt-2" style={styles.subtitleFont}>
-        A quiet space for intentional readers
-      </Text>
-
-      <View className="mt-10 gap-5">
-        {error && <Text className="text-primary text-sm text-center">{error}</Text>}
-
-        <View className="gap-1.5">
-          <Label htmlFor="name" className={labelClassName}>
-            Name
-          </Label>
-          <Input
-            id="name"
-            placeholder="Your full name"
-            value={name}
-            onChangeText={setName}
-            className={inputClassName}
-          />
-          {fieldErrors.name?.map((e) => (
-            <Text key={e} className="text-primary text-xs">
-              {e}
-            </Text>
-          ))}
-        </View>
-
-        <View className="gap-1.5">
-          <Label htmlFor="email" className={labelClassName}>
-            Email
-          </Label>
-          <Input
-            id="email"
-            placeholder="you@example.com"
-            value={email}
-            onChangeText={setEmail}
-            autoCapitalize="none"
-            keyboardType="email-address"
-            className={inputClassName}
-          />
-          {fieldErrors.email_address?.map((e) => (
-            <Text key={e} className="text-primary text-xs">
-              {e}
-            </Text>
-          ))}
-        </View>
-
-        <View className="gap-1.5">
-          <Label htmlFor="password" className={labelClassName}>
-            Password
-          </Label>
-          <Input
-            id="password"
-            placeholder="••••••••"
-            value={password}
-            onChangeText={setPassword}
-            secureTextEntry
-            className={inputClassName}
-          />
-          {fieldErrors.password?.map((e) => (
-            <Text key={e} className="text-primary text-xs">
-              {e}
-            </Text>
-          ))}
-        </View>
-
-        <View className="gap-1.5">
-          <Label htmlFor="confirm" className={labelClassName}>
-            Confirm Password
-          </Label>
-          <Input
-            id="confirm"
-            placeholder="••••••••"
-            value={confirm}
-            onChangeText={setConfirm}
-            secureTextEntry
-            className={inputClassName}
-          />
-        </View>
-
-        <Button
-          className="w-full bg-primary rounded-sm mt-2"
-          onPress={handleSignUp}
-          disabled={loading}
-        >
-          {loading ? (
-            <ActivityIndicator color={colors.onPrimary} />
-          ) : (
-            <Text className="text-on-primary text-xs font-medium uppercase tracking-wider text-center">
-              Create Account
-            </Text>
-          )}
-        </Button>
-      </View>
-
-      <Text className="text-center text-sm text-outline mt-6">
-        Already have an account?{" "}
-        <Text
-          className="text-primary text-sm font-medium"
-          onPress={() => router.push("/(auth)/sign-in")}
-        >
-          Sign in
-        </Text>
-      </Text>
-    </Screen>
+    <SignUpScreen
+      loading={loading}
+      error={error}
+      fieldErrors={fieldErrors}
+      onSignUp={handleSignUp}
+      onSignIn={() => router.push("/(auth)/sign-in")}
+    />
   )
 }
-
-const styles = StyleSheet.create({
-  content: {
-    paddingBottom: 32,
-    paddingHorizontal: 24,
-  },
-  subtitleFont: {
-    fontFamily: "newsreaderRegularItalic",
-  },
-  titleFont: {
-    fontFamily: "newsreaderMedium",
-  },
-})
