+++ src/components/Auth/authFormStyles.ts
@@ -1,0 +1,4 @@
+export const inputClassName =
+  "bg-surface-container border-outline-variant/50 rounded-sm h-[50px] px-[14px] text-sm text-on-surface"
+
+export const labelClassName = "text-xs uppercase tracking-[0.1em] text-on-surface"
